// src/models/LiveStreaming/SessionInfo.ts
export interface SessionInfo {
    Id: string;
    LiveStreamingId: string;
    SessionName: string;
    SessionStartTime: string | null;
    SessionEndTime: string | null;
    VAExposureStartTime: string | null;
    VAExposureEndTime: string | null;
    Created: string | null;
    Updated: string | null;
}